requirements/requirements.txt
asgiref==3.3.4
Django==3.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
